udated admin/rc/proforma/:id
